requirements.txt
streamlit==1.37.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
pymupdf==1.24.7
langchain-text-splitters==0.2.2
numpy==1.26.4
torch>=2.2.0
transformers>=4.41.0
